refactor: split toolchain inspect modules
